requirements.txt
requests
jsonschema>=3.0.0
confluent_kafka
pytimeseries
pyyaml
